修复的问题 1. UPnP使用NIO 2. 让UDP Channel更易于使用 3. SplittedSource现已支持不规则的文件大小 4. Compiler的输出没有颜色 5. AnsiString的分行方式有误 6. XML TextNode没有正确设置自身的类型 7. XML Node QuerySelector不能使用 8. MyHashMap在上次树化更新后hash计算错误 9. 为不支持ANSI转义的终端加入了兼容层 10. 修改了CommandConsole的dumpNodes 11. UnsafeCharset
新增的功能
1. 为LZMA(2)InputStream实现了skip
2. BufferPool可以缩小缓冲区大小
3. 为DefaultConsole加入了异步可用的readLine和关闭回显等函数

+++ roj/reflect/EnumHelper.java
@@ -71,8 +71,8 @@ public final class EnumHelper extends CodeVisitor {
 
 					long v = 0;
 					v |= (long)i << 24;
-					v |= nameId << 16;
-					v |= ordinalId << 8;
+					v |= (long) nameId << 16;
+					v |= (long) ordinalId << 8;
 					v |= param.size();
 					parPos.put(TypeHelper.getMethod(param), v);
 				}
